refactor: rename npx module to use models for data representation, matching thirteenf conventions
